requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
Werkzeug==3.1.3
